Proposal enable css modules filename convention
Hanna, Ollie, Alistair, Tim and I went over your proposal together to discuss it.

We’ve spent a lot of time going over this and we try to accept proposals where possible but we have decided not to accept this proposal for now for the following reasons:

1. We are not confident that we would be able to support this when issues come up.
2. We don’t currently have a way to publish ‘alpha’ / ‘experimental’ features without making breaking changes for all users.
3. We don't have time - due to constraints outside our control we need to be stricter with what we can do and can’t prioritise this work.

In order to accept this in the future we’d need:
1. Evidence that CSS Modules work for all components, and how they are tested.
2. Evidence of the need from teams across Government
3. More time as a team

We think that the underlying problem this tries to solve is a multi-package distribution of components. We are not not sure this has enough benefits to warrant the extra overhead required to support this, but you could consider publishing govuk-react as a single package.

You can read more about our decision to publish a single package here: https://github.com/alphagov/govuk-design-system-architecture/blob/master/proposals/002-publish-one-npm-package-instead-of-multiple-npm-packages.md

This said, we still believe that more investigation into CSS Modules is a worthwhile thing in the future, and hope that you continue to explore it separately within your experimental package.

We’d also be open to documenting an experimental CSS Modules package within the GOV.UK Frontend installation documentation so that we can build more confidence.

Co-authored-by: Hanna Laakso <hanna.laakso@digital.cabinet-office.gov.uk>
Co-authored-by: Alasdair McLeay <penx@users.noreply.github.com>